@@ -377,15 +377,17 @@ class Validate
      * @param string $scene 验证场景
      * @return bool
      */
-    public function check($data, $rules = [], $scene = null)
+    public function check($data, $rules = [], $scene = '')
     {
         $this->error = [];
+
         if (empty($rules)) {
             // 读取验证规则
-            $rules = $this->getSceneRule($scene ?? $this->currentScene);
+            $rules = $this->rule;
         }
 
         // 获取场景定义
+        $this->getScene($scene);
 
         foreach ($this->append as $key => $rule) {
             if (!isset($rules[$key])) {
@@ -394,13 +396,20 @@ class Validate
         }
 
         foreach ($rules as $key => $rule) {
+
             // field => 'rule1|rule2...' field => ['rule1','rule2',...]
+            if (is_numeric($key)) {
+                $key = $rule;
+                $rule = $this->rule[$key] ?? '';
+            }
             if (is_array($rule)) {
                 $rule = array_filter($rule);
             }
+
             if (empty($rule)) {
                 continue;
             }
+
             // field => 'rule1|rule2...' field => ['rule1','rule2',...]
             if (strpos($key, '|')) {
                 // 字段|描述 用于指定属性名称
@@ -409,7 +418,7 @@ class Validate
                 $title = isset($this->field[$key]) ? $this->field[$key] : $key;
             }
 
-            // 获取数据 支持多维数组
+            // 获取数据 支持二维数组
             $value = $this->getDataValue($data, $key);
             switch (true) {
                 case $rule instanceof \Closure:
@@ -422,29 +431,23 @@ class Validate
                     $result = $this->checkItem($key, $value, $rule, $data, $title);
                     break;
                 case is_array($rule):
-                    if (!is_array($value)) {
+                    if (!isset($data[$key]) || !is_array($data[$key])) {
                         $result = '参数' . $key . "必须为二维数组";
                         break;
                     }
                     $ruleStr = [];
                     foreach ($rule as $ruleKey => $itemRule) {
-                        $ruleStr[$ruleKey] = $itemRule;
+                        $field = str_replace('*.', '', $ruleKey);
+                        $ruleStr[$field] = $itemRule;
                     }
-                    $iv = 0;
-                    $newVal = new Validate([], [], $this->field[$key] ?? []);
-                    foreach ($value as $item) {
-                        $iv++;
+                    foreach ($data[$key] as $item) {
                         if (!is_array($item)) {
-                            $result = $key . "第{$iv}项必须为二维数组";
+                            $result = $key . "必须为二维数组";
                             break;
                         }
-                        $result = $newVal->check($item, $ruleStr, $key);
-                        if ($result !== true) {
-                            $this->error = $key . "第{$iv}项的" . $newVal->getError();
-                            return false;
-                        }
+                        $result = $this->check($item, $ruleStr, $key);
+                        if ($result !== true) return false;
                     }
-                    return true;
                     break;
                 default:
                     $result = $this->checkItem($key, $value, $rule, $data, $title);
@@ -465,6 +468,7 @@ class Validate
                 }
             }
         }
+
         return !empty($this->error) ? false : true;
     }
 
@@ -573,6 +577,7 @@ class Validate
             } else {
                 // 判断验证类型
                 list($type, $rule, $info) = $this->getValidateType($key, $rule);
+
                 if (isset($this->remove[$field]) && in_array($info, $this->remove[$field])) {
                     // 规则已经移除
                     $i++;
@@ -1400,15 +1405,13 @@ class Validate
         if (is_numeric($key)) {
             $value = $key;
         } elseif (strpos($key, '.')) {
-            // 支持多维数组验证
-            $tmp = $data;
-            foreach (explode('.', $key) as $segment) {
-                $tmp = $tmp[$segment] ?? null;
-            }
-            $value = $tmp;
+            // 支持二维数组验证
+            list($name1, $name2) = explode('.', $key);
+            $value = isset($data[$name1][$name2]) ? $data[$name1][$name2] : null;
         } else {
             $value = isset($data[$key]) ? $data[$key] : null;
         }
+
         return $value;
     }
 
@@ -1504,30 +1507,9 @@ class Validate
         return call_user_func_array([$this, 'is'], $args);
     }
 
-    public function getSceneRule( $name)
+    public function getSceneRule(string $name)
     {
-        $rules = $this->scene[$name] ?? $this->rule;
-        $newRules = [];
-        foreach ($rules as $key => $rule) {
-            if (is_numeric($key)) {
-                $key = $rule;
-                $rule = $this->rule[$key] ?? '';
-            }
-            if (!empty($rule)) $newRules[$key] = $rule;
-            $this->getRule($key, $newRules);
-        }
-        return $newRules;
-    }
-
-
-    public function getRule($field, &$newRules)
-    {
-        foreach ($this->rule as $key => $rule) {
-            if (strpos($key, $field . ".") !== false) {
-                $newRules[$key] = $rule;
-            }
-        }
-        return $newRules;
+        return $this->scene[$name] ?? $this->rule;
     }
 
 }
